fix: вынес dto в отдельный проект, чтобы можно было использовать и на сервере, и на клиенте
