+++ DispatchTransportControl.Api/Mapping/MappingProfile.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
-using DispatchTransportControl.Api.DTO;
 using DispatchTransportControl.Domain.Entity;
+using DispatchTransportControl.Shared.DTO;
 
 namespace DispatchTransportControl.Api.Mapping;
