+++ DispatchTransportControl.Domain/DispatchTransportControl.Domain.csproj
@@ -13,4 +13,8 @@
         <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.0"/>
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\DispatchTransportControl.Shared\DispatchTransportControl.Shared.csproj" />
+    </ItemGroup>
+
 </Project>
