+++ DispatchTransportControl.Shared/DispatchTransportControl.Shared.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <RootNamespace>DispatchTransportControl.Shared</RootNamespace>
+    </PropertyGroup>
+
+</Project>
