+++ DispatchTransportControl.Shared/VehicleType.cs
@@ -1,0 +1,11 @@
+﻿namespace DispatchTransportControl.Shared;
+
+/// <summary>
+///     Типы транспортных средств
+/// </summary>
+public enum VehicleType
+{
+    Bus,
+    Trolley,
+    Tram
+}
